Changes to be committed: 	modified:   ex00/BitcoinExchange.cpp 	modified:   ex00/BitcoinExchange.hpp 	modified:   ex00/Makefile 	new file:   ex00/input 	deleted:    ex00/input.csv 	modified:   ex00/main.cpp 	modified:   ex01/Makefile 	modified:   ex01/RPN.cpp

+++ ex00/BitcoinExchange.hpp
@@ -50,7 +50,7 @@ class BitcoinExchange {
 	double		validate(int i);
 	int			findPrevious(int i);
 	int			find(int i);
-	void		print(int i, int value, int index);
+	void		print(int i, double value, int index);
 	void		print();
 
 	class OpenFile : public std::exception {
